First Official Release Commit :D
I finally got the thing to work :)

+++ src/com/lh64/gltextures/SmartTexture.java
@@ -93,7 +93,7 @@ public class SmartTexture extends Texture {
 		bitmap = null;
 	}
 	
-	public RectF uvRect( int left, int top, int right, int bottom ) {
+	public RectF uvRect( int left, int top, float right, float bottom ) {
 		return new RectF(
 			(float)left		/ width,
 			(float)top		/ height,
